moving params to a different file
